re/mergeTest
드가자

+++ Assets/01_Scripts/EnvScripts/SimpleFloatings.cs
@@ -6,17 +6,17 @@ using DG.Tweening;
 public class SimpleFloatings : MonoBehaviour
 {
     private Transform floatingTr;
-    float rnd;
+    int rnd;
 
     void Start()
     {
         DOTween.Init();
         floatingTr = GetComponent<Transform>();
 
-        rnd = Random.Range(3, 10f);
+        rnd = Random.Range(3, 10);
 
         Sequence seq = DOTween.Sequence();
-        seq.Append(floatingTr.DOMoveY(floatingTr.position.y + rnd / 15, rnd)).Append(floatingTr.DOMoveY(floatingTr.position.y, rnd).OnComplete<Tween>(() => rnd = Random.Range(3, 10f)))
+        seq.Append(floatingTr.DOMoveY(floatingTr.position.y + rnd / 15, rnd)).Append(floatingTr.DOMoveY(floatingTr.position.y, rnd).OnComplete<Tween>(() => rnd = Random.Range(3, 10)))
             .SetLoops<Sequence>(-1);
 
         //floatingTr.DORotate()
